Adds working qrcode generation as guid token not ut123456 code (better security) to change variable names thats rly mess up

+++ QrToPay/View/SkiResort/SkiResortCityPage.xaml.cs
@@ -14,6 +14,6 @@ public partial class SkiResortCityPage : ContentPage
 
     private async void InitializeAsync()
     {
-        await _viewModel.LoadCitiesAsync(EntityCategory.SkiResort);
+        await _viewModel.LoadCitiesAsync((int)EntityCategory.SkiResort);
     }
 }
